Generalize Highlightmode visible highligts
In the past, setting the visible highlights was made with adding/removing a markerClass in the wrapper element.
This does not work for image annotations so the setVisibleHighlights event was introduced. However textHighlights did not use that, and we manually switched the markerClass on/off.
Now texthighlights switches this class reacting to the event.

Imageanchors now listens to the annotationsLoaded event and switches on the visibleHighlight mode when it is needed.

Fixes #987