Add docker-compose.yml for building dev environ
This adds a docker-compose.yml, used (unsurprisingly) by
docker-compose[1] to assemble an application out of multiple pieces.
This configuration sets up a basic development environment for h, with
the following services:

- web (the application)
- elasticsearch
- nsqd

At the moment postgres is not configured, and the application will
default to using an SQLite database within the container.

The simplest way to run this all at the moment is probably to start the
services in the background first:

    docker-compose up -d elasticsearch nsqd

And then start the web application without recreating the dependency
containers. This is to work around docker-compose issue #919[2].

    docker-compose up --no-deps --no-recreate web

This will install development dependencies and start the web
application. You can then visit the application at port 8000 on the
docker host.

[1]: https://docs.docker.com/compose/
[2]: https://github.com/docker/compose/issues/919